+++ Sources/APRelay/App/routes.swift
@@ -1,4 +1,3 @@
-import Prometheus
 import Vapor
 
 func routes(_ app: Application) throws {
@@ -8,9 +7,4 @@ func routes(_ app: Application) throws {
     try app.register(collection: WebFingerController())
     try app.register(collection: NodeInfoController())
     try app.register(collection: AdminAPIController())
-
-    // Prometheus metrics endpoint.
-    app.get("metrics") { req -> String in
-        req.application.prometheusRegistry?.emitToString() ?? ""
-    }
 }